services/trending-service/requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
python-dotenv==1.0.0
google-search-results==2.4.2
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
scikit-learn==1.3.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9 